isolate server creation
